# perf(stdlib): optimize window transformation