… stdlib/universe/window.go