This optimizes the window transformation to be more efficient than the 
current window transformation. It changes the access pattern so the time 
values are iterated over in sorted order and the windows are created in 
sorted order. This reduces the amount of random access used when iterating
over values so memory is more efficient.

It also adds a new `RandomAccessGroupLookup` that uses another method for
storing values based on their group key. The window method had a bad access
pattern for the standard `GroupLookup` because the `_start` and
`_stop` values changing would commonly result in an access pattern like 
this:

    1 2 3 4 5 6 -> 1 4 2 5 3 6

This hit one of the group lookups worst access patterns and caused a lot of
time to be spent copying memory to maintain the sorted order of the data
structure.

The window transformation still likely has more room to improve, but this
will at least cause some small amount of improvement over the old 
implementation.

### Done checklist
- [x] docs/SPEC.md updated
- [x] Test cases written