@jsternberg 2020-01-02
Benchmarks:

```
benchmark                            old ns/op        new ns/op      delta
BenchmarkWindow_Infinity_1000-12     105918759196     796584582      -99.25%
BenchmarkWindow_ByMinute_100-12      60009869814      4825699308     -91.96%

benchmark                            old allocs     new allocs     delta
BenchmarkWindow_Infinity_1000-12     696462018      1302642        -99.81%
BenchmarkWindow_ByMinute_100-12      152100590      43378993       -71.48%

benchmark                            old bytes       new bytes      delta
BenchmarkWindow_Infinity_1000-12     33036214744     576351637      -98.26%
BenchmarkWindow_ByMinute_100-12      8735916472      3785190209     -56.67%
```

The infinity implementation is a lot better and should help `aggregateWindow`. The normal implementation isn't as good and will likely be the target of future improvements, but it was good enough that I didn't want to delay good just because it wasn't perfect.